double values

+++ Microsoft.OpenAIRateLimiter.API/QuotaRequest.cs
@@ -64,7 +64,7 @@ namespace Microsoft.OpenAIRateLimiter.API
 
                 return HttpUtilities.RESTResponse(await _svc.Create(new QuotaDTO() { Key = data.SubscriptionKey, 
                                                                                      Product = data.ProductName,
-                                                                                     Value = Convert.ToInt32(data.Amount) }));
+                                                                                     Value = Convert.ToDouble(data.Amount) }));
 
             }
             catch (Exception ex)
@@ -133,7 +133,7 @@ namespace Microsoft.OpenAIRateLimiter.API
                     return HttpUtilities.RESTResponse(alert?.data?.alertContext?.AlertData?.BudgetName);
                 }
 
-                return HttpUtilities.RESTResponse(await _svc.BudgetUpdate(new QuotaDTO() { Product = alert.data.alertContext.AlertData.BudgetName, Value = 0 }));
+                return HttpUtilities.RESTResponse(await _svc.BudgetUpdate(new QuotaDTO() { Product = alert.data.alertContext.AlertData.BudgetName, Value = 0.00 }));
 
             }
             catch (Exception ex)
